requirements.txt
# requirements file
crewai
crewai_tools
python-dotenv